@@ -18,8 +18,8 @@
  */
 package org.apache.cxf.fediz.core.handler;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 public interface RequestHandler<T> {
 
