@@ -21,7 +21,7 @@ package org.apache.cxf.fediz.core.spi;
 
 import java.util.regex.Pattern;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 public class ReplyConstraintCallback extends AbstractServletCallback {
 
