@@ -19,9 +19,8 @@
 
 package org.apache.cxf.fediz.tomcat.handler;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.apache.catalina.Session;
 import org.apache.catalina.connector.Request;
 import org.apache.cxf.fediz.core.config.FedizContext;
